Move all packages into cvb package
