+++ src/main/java/cvb/commands/CommandParser.java
@@ -1,14 +1,14 @@
-package commands;
+package cvb.commands;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.stream.Collectors;
 
-import exceptions.ConvoBotException;
-import tasks.Deadline;
-import tasks.Event;
-import tasks.ToDo;
-import utils.DateTime;
+import cvb.exceptions.ConvoBotException;
+import cvb.tasks.Deadline;
+import cvb.tasks.Event;
+import cvb.tasks.ToDo;
+import cvb.utils.DateTime;
 
 /**
  * The {@code CommandParser} class provides utility methods for parsing commands from user input.
