+++ src/main/java/cvb/gui/DialogBox.java
@@ -1,4 +1,4 @@
-package gui;
+package cvb.gui;
 
 import java.io.IOException;
 import java.util.Collections;
