+++ src/main/java/cvb/utils/DateTime.java
@@ -1,10 +1,10 @@
-package utils;
+package cvb.utils;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-import exceptions.ConvoBotException;
+import cvb.exceptions.ConvoBotException;
 
 /**
  * The {@code DateTime} class provides utility methods for handling date and time operations.
