+++ src/test/java/convobot/tasks/TaskParserTest.java
@@ -1,4 +1,4 @@
-package tasks;
+package cvb.tasks;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
